+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
+### Updated
+- Updated DOM content structure of `Modal` component
 
 ## [0.17.0] - 2019-09-10
 ### Added